Enhancing system scalbility, creating reusable functions and updating the verse of the day every 12 midnight
