+++ src/hooks/verseOfTheDay.js
@@ -1,14 +1,15 @@
 import axios from 'axios';
 import * as cheerio from 'cheerio';
 
-export async function getVerseOfTheDay() {
+export async function getVerseOfTheDayWithImage() {
   const URL = 'https://www.bible.com/verse-of-the-day';
   try {
     const {data} = await axios.get(URL);
     const $ = cheerio.load(data);
 
-    const versesArray: Array<String> = [];
-    const citationsArray: Array<String> = [];
+    const versesArray = [];
+    const citationsArray = [];
+    const imageArray = [];
     const verses = $('p.text-gray-50');
     const citations = $('.mbs-2');
 
@@ -23,9 +24,13 @@ export async function getVerseOfTheDay() {
       versesArray.push(formattedVerse);
     });
 
+    const imageUrl = $('img').attr('src');
+    imageArray.push(imageUrl);
+
     return {
       citation: citationsArray[1],
       passage: versesArray[0],
+      imageUrl: imageArray[0],
     };
   } catch (err) {
     console.error(err);
